Create beat the clock
Sample code for beat-the-clock gameplay with 3 second countdown timer. Score is x3 if player matches correctly in 1s or less, score x2 if player matches in less than 2 seconds, score is 1 point if player matches in less than 3 seconds. Score is zero if time expires. When time expires, load new image and reset timer with same rules as above.